Only run certificate exclusion once
This command will fail if ran a second time, so only run it if the
certificate has already been "fixed".

+++ recipes/default.rb
@@ -31,7 +31,7 @@ end
 # https://blog.devgenius.io/rhel-centos-7-fix-for-lets-encrypt-change-8af2de587fe4
 execute "fix certificates" do
   command 'trust dump --filter "pkcs11:id=%c4%a7%b1%a4%7b%2c%71%fa%db%e1%4b%90%75%ff%c4%15%60%85%89%10" | openssl x509 | sudo tee /etc/pki/ca-trust/source/blacklist/DST-Root-CA-X3.pem'
-  ignore_failure true
+  not_if { ::File.exists?("/etc/pki/ca-trust/source/blacklist/DST-Root-CA-X3.pem") }
 end
 
 execute "update root store" do
